refactor(ui): validation 누락된 부분 추가
