+++ src/main/java/kitchenpos/ui/dto/OrderTableIdRequest.java
@@ -1,7 +1,10 @@
 package kitchenpos.ui.dto;
 
+import javax.validation.constraints.NotNull;
+
 public class OrderTableIdRequest {
 
+    @NotNull
     private Long id;
 
     public OrderTableIdRequest() {
