+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -7,7 +7,7 @@ import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.order.OrderTable;
 import kitchenpos.domain.order.TableGroup;
 import kitchenpos.domain.product.Product;
-import kitchenpos.ui.OrderTableResponse;
+import kitchenpos.ui.dto.OrderTableResponse;
 import kitchenpos.ui.dto.OrderTableEmptyRequest;
 import kitchenpos.ui.dto.OrderTableGuestRequest;
 import kitchenpos.ui.dto.OrderTableRequest;
